Projektmappen richtig erstellt
