--- Game-Of-Life-Linq.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.11.35303.130
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Game-Of-Life-Linq", "Game-Of-Life-Linq.csproj", "{E24896ED-E57E-41E6-830F-23AC4AE8E48C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E24896ED-E57E-41E6-830F-23AC4AE8E48C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E24896ED-E57E-41E6-830F-23AC4AE8E48C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E24896ED-E57E-41E6-830F-23AC4AE8E48C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E24896ED-E57E-41E6-830F-23AC4AE8E48C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1C52F20B-2B0A-48AF-AE36-95C8552D73E5}
-	EndGlobalSection
-EndGlobal
