+++ src/GOL.Business/GameEngine.cs
