+++ src/GOL.Console/GOL.Console.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\GOL.Business\GOL.Business.csproj" />
+    <ProjectReference Include="..\GOL.Contract\GOL.Contract.csproj" />
+  </ItemGroup>
+
+</Project>
