+++ src/GOL.Contract/Cell.cs
