+++ src/GOL.Contract/IGameEngine.cs
